worker/requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
requests-mock>=1.11.0
